Finally, killed old MC::mergeLogFiles code We now call vk_logmerge via a VKProcess Using same process framework as MC::run()  - this code needed only a little work to generalise it.
TODO
A side effect of this is that MC::merge will now listen to the flags '--log-file', and '--log-file-exactly'.  This is a 'good thing'(tm) in that users can specify the output of the merge nice and easily.  However, it's not documented behaviour, since those options appear to be only applicable to a valgrind run.  This needs thinking about for a proper solution.